refactor: test.service.member에서 있던 클래스들을 MemberServiceTest로 통합
